Baseline for load songs feature and some refactoring again
