+++ swiftify-frontend/src/interface_adapter/load_songs/LoadSongsController.java
@@ -1,0 +1,16 @@
+package interface_adapter.load_songs;
+
+import use_case.load_songs.LoadSongsInputBoundary;
+
+public class LoadSongsController {
+
+    final LoadSongsInputBoundary loadSongsInputBoundary;
+
+    public LoadSongsController(LoadSongsInputBoundary loadSongsInputBoundary) {
+        this.loadSongsInputBoundary = loadSongsInputBoundary;
+    }
+
+    public void execute() {
+        loadSongsInputBoundary.execute();
+    }
+}
