--- swiftify-frontend/src/use_case/display_albums/DisplayAlbumsDataAccessInterface.java
@@ -1,8 +1,0 @@
-package src.use_case.display_albums;
-
-import java.util.List;
-
-public interface DisplayAlbumsDataAccessInterface {
-
-    public List<String> getAlbumNames();
-}
