--- swiftify-frontend/src/use_case/display_albums/DisplayAlbumsInputBoundary.java
@@ -1,6 +1,0 @@
-package src.use_case.display_albums;
-
-public interface DisplayAlbumsInputBoundary {
-
-    public void execute();
-}
