--- swiftify-frontend/src/use_case/display_albums/DisplayAlbumsInteractor.java
@@ -1,13 +1,0 @@
-package src.use_case.display_albums;
-
-public class DisplayAlbumsInteractor implements DisplayAlbumsInputBoundary {
-
-    public DisplayAlbumsInteractor() {
-
-    }
-
-    @Override
-    public void execute() {
-        // query S3 database to get list of album names, then present back to view
-    }
-}
