+++ swiftify-frontend/src/use_case/load_songs/LoadSongsDataAccessInterface.java
@@ -1,0 +1,8 @@
+package use_case.load_songs;
+
+import java.util.List;
+
+public interface LoadSongsDataAccessInterface {
+
+    public List<String> getAlbumNames();
+}
