+++ swiftify-frontend/src/use_case/load_songs/LoadSongsInputBoundary.java
@@ -1,0 +1,6 @@
+package use_case.load_songs;
+
+public interface LoadSongsInputBoundary {
+
+    public void execute();
+}
