+++ swiftify-frontend/src/use_case/load_songs/LoadSongsOutputData.java
@@ -1,13 +1,13 @@
-package src.use_case.display_albums;
+package use_case.load_songs;
 
 import java.util.List;
 
-public class DisplayAlbumsOutputData {
+public class LoadSongsOutputData {
 
     private final List<String> albumNames;
     private boolean useCaseFailed;
 
-    public DisplayAlbumsOutputData(List<String> albumNames, boolean useCaseFailed) {
+    public LoadSongsOutputData(List<String> albumNames, boolean useCaseFailed) {
         this.albumNames = albumNames;
         this.useCaseFailed = useCaseFailed;
     }
