+++ swiftify-frontend/src/use_case/next_song/NextInputBoundary.java
@@ -1,4 +1,4 @@
-package src.use_case.next_song;
+package use_case.next_song;
 
 public interface NextInputBoundary {
     void execute(NextInputData nextInputData);
