+++ swiftify-frontend/src/use_case/next_song/NextInteractor.java
@@ -1,4 +1,4 @@
-package src.use_case.next_song;
+package use_case.next_song;
 
 public class NextInteractor implements NextInputBoundary {
     final NextOutputBoundary nextOutputBoundary;
