+++ swiftify-frontend/src/use_case/next_song/NextOutputBoundary.java
@@ -1,4 +1,4 @@
-package src.use_case.next_song;
+package use_case.next_song;
 
 public interface NextOutputBoundary {
 }
