+++ swiftify-frontend/src/use_case/next_song/NextOutputData.java
@@ -1,4 +1,4 @@
-package src.use_case.next_song;
+package use_case.next_song;
 
 import java.util.Stack;
