+++ swiftify-frontend/src/use_case/previous_song/PreviousInputBoundary.java
@@ -1,4 +1,4 @@
-package src.use_case.previous_song;
+package use_case.previous_song;
 
 public interface PreviousInputBoundary {
     void execute(PreviousInputData nextInputData);
