+++ swiftify-frontend/src/use_case/previous_song/PreviousInteractor.java
@@ -1,4 +1,4 @@
-package src.use_case.previous_song;
+package use_case.previous_song;
 
 public class PreviousInteractor implements PreviousInputBoundary {
     final PreviousOutputBoundary previousOutputBoundary;
