+++ swiftify-frontend/src/use_case/previous_song/PreviousOutputData.java
@@ -1,4 +1,4 @@
-package src.use_case.previous_song;
+package use_case.previous_song;
 
 import java.util.Stack;
